fix platform/aws storage; remove bucket trait
